feat: DataFrame pickle support for Dask, Ray and multiprocessing.
This makes it easier to pass around a dataframe in Dask, Ray or using
multiprocessing.

Note that hdf5 or parquet files will be pickled really efficiently,
since it's the *path* that is passed around, not the data.

For Arrow IPC files we do pass around the data. We should change/refactor
that.

Also, CSV files lose their origin, so all the data will be transported.

Note that this also makes it really easy to store your whole 'session',
by pickling the dataframe to disk.