fix: wrap arrays returned from the selection scope
Fixed #1111

Otherwise NumPy can do things elementwise.
Note that variables are not wrapped, otherwise it's a bit hard to use
when variables are used in lists (which are not inspected).

On the long term we may want to rewrite the AST to use functions, instead
of overriding the `__mul__` operators etc.