✨ file stream object to allow writing to a file and consuming chunks
this is useful in combination with e.g. FastAPI where you want to
asynchronously consume and write chunks of data to the socket.

Writing should be done from a thread, and back pressure is provided
by using a threading.Queue between the writer and consumer.

An async generator makes use or starlette (not a depency of vaex-core
but of vaex-server via fastapi)