feat: Add cmake & ninja build requirements only when needed
Rather than listing `cmake` and `ninja` build requirements
unconditionally, add them to `setup_requires` when they are necessary --
that is, when using skbuild and when the system installations of these
tools are not available.  This avoids installing the packages
unnecessarily, and improves portability by using downstream patched
CMake version when available.